refactor(auth): 인증 코드를 한곳에 모으고 서버 전용 코드를 분리
