+++ apps/web/src/modules/auth/index.ts
@@ -1,3 +1,4 @@
+export { safeCallbackUrl } from "./callback-url";
 export { LoginForm } from "./components/login-form";
 export { SignOutButton } from "./components/sign-out-button";
 export type { LoginInput, RegisterInput } from "./schemas/auth.schema";
