+++ apps/web/src/modules/auth/server/auth.config.ts
@@ -1,0 +1,23 @@
+import type { NextAuthConfig } from "next-auth";
+
+export const authConfig: NextAuthConfig = {
+	pages: {
+		signIn: "/login",
+		error: "/login",
+	},
+	callbacks: {
+		jwt({ token, user }) {
+			if (user) {
+				token.id = user.id;
+			}
+			return token;
+		},
+		session({ session, token }) {
+			if (token && session.user) {
+				session.user.id = token.id as string;
+			}
+			return session;
+		},
+	},
+	providers: [],
+};
