+++ apps/web/src/modules/auth/server/index.ts
@@ -1,2 +1,4 @@
+import "server-only";
+
 export { auth, handlers, signIn, signOut } from "./auth";
 export { authConfig } from "./auth.config";
